refactor stepexecution
