--- src/main/java/executor/service/service/stepexecutors/selenium/impl/StepExecutionClickXpathImpl.java
@@ -1,17 +1,0 @@
-package executor.service.service.stepexecutors.selenium.impl;
-
-import executor.service.model.entity.Step;
-import executor.service.service.stepexecutors.selenium.StepExecutionClickXpath;
-import org.openqa.selenium.WebDriver;
-
-public class StepExecutionClickXpathImpl implements StepExecutionClickXpath {
-    @Override
-    public String getStepAction() {
-        return null;
-    }
-
-    @Override
-    public void step(WebDriver webDriver, Step step) {
-
-    }
-}
